workplacesafety/requirements.txt
azureml-sdk[notebooks, automl]
numpy
Pillow
protobuf
tensorflow==1.10.0
jupyter
matplotlib